#21 make load.py on model/tokenizer
